Commit con la mayoria de los servicios funcionando salvo update

+++ backend/src/main/java/com/portafolio/backend/Repository/EducacionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducacionRepository extends JpaRepository<EducacionModel, Long> {
+public interface EducacionRepository extends JpaRepository<EducacionModel, Integer> {
 
 }
