+++ backend/src/main/java/com/portafolio/backend/Repository/ExperienciaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienciaRepository extends JpaRepository<ExperienciaModel, Long> {
+public interface ExperienciaRepository extends JpaRepository<ExperienciaModel, Integer> {
 
 }
