+++ backend/src/main/java/com/portafolio/backend/Repository/HabilidadesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HabilidadesRepository extends JpaRepository<HabilidadesModel, Long> {
+public interface HabilidadesRepository extends JpaRepository<HabilidadesModel, Integer> {
 
 }
